-overwrite: seek instead of re-reading partition table
The upside is that this fixes setups where $GOBIN resides in a file system which
is mounted nosuid (where capabilities are not applied), because we’ll do away
with the setcap call in favor of a setfacl call.

The downside is that the SD card needs to be unplugged and re-plugged before the
persistent file system can be created (if desired).